backend/utils/supabase_client.py
import os
import streamlit as st
from supabase import create_client, Client
from dotenv import load_dotenv

load_dotenv()

class SupabaseClient:
    def __init__(self):
        self.url = os.getenv("SUPABASE_URL")
        self.key = os.getenv("SUPABASE_KEY")
        self.client: Client = None
        
        if self.url and self.key:
            try:
                self.client = create_client(self.url, self.key)
            except Exception as e:
                st.error(f"❌ Supabase connection error: {e}")
    
    def is_connected(self):
        return self.client is not None

# Global instance
supabase = SupabaseClient()