requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1

# AI and Language Models
langchain==0.1.0
langchain-google-genai==0.0.6
google-generativeai==0.3.2

# Data Processing
pandas==2.1.3
plotly==5.17.0

# Database and Storage
supabase==2.3.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
